+++ changelog.md
@@ -8,7 +8,9 @@ This version is not yet released. If you are reading this on the website, then t
 - **Breaking Change** - [`un °`](https://uiua.org/docs/un) [`fix ¤`](https://uiua.org/docs/fix) now does pattern matching
   - [`under ⍜`](https://uiua.org/docs/under) [`fix ¤`](https://uiua.org/docs/fix) retains the old behavior
   - Collapsing the top two dimensions of an array can still be done with [`reduce /`](https://uiua.org/docs/reduce)[`join ⊂`](https://uiua.org/docs/join)
-- **Breaking Change** - [`un °`](https://uiua.org/docs/un) [`on ⟜`](https://uiua.org/docs/on) now does pattern matching
+- **Breaking Change** - [`un °`](https://uiua.org/docs/un) [`on ⟜`](https://uiua.org/docs/on) now does pattern matching in some cases
+  - This makes its bevahior conform to the rule that a function's inverse must have the opposite signature
+  - Some other cases, such as `°⟜+`, now work as you might expect
 - **Breaking Change** - [`keep ▽`](https://uiua.org/docs/keep) with a scalar counts array now copies each row that many times
   - This matches the behaviors of APL and BQN, and is generally more useful
   - The old behavior can be achieved with `/⊂↯`